+++ packages/wonder-blocks-form/src/components/labeled-text-field.tsx
@@ -30,7 +30,15 @@ type CommonProps = {
      */
     value: string;
     /**
-     * Makes a read-only input field that cannot be focused. Defaults to false.
+     * Whether the input should be disabled. Defaults to false.
+     * If the disabled prop is set to `true`, LabeledTextField will have disabled
+     * styling and will not be interactable.
+     *
+     * Note: The `disabled` prop sets the `aria-disabled` attribute to `true`
+     * instead of setting the `disabled` attribute. This is so that the component
+     * remains focusable while communicating to screen readers that it is disabled.
+     * This `disabled` prop will also set the `readonly` attribute to prevent
+     * typing in the field.
      */
     disabled: boolean;
     /**
@@ -118,6 +126,10 @@ type CommonProps = {
      * Specifies if the TextField allows autocomplete.
      */
     autoComplete?: string;
+    /**
+     * Provide a name for the TextField.
+     */
+    name?: string;
 };
 
 type OtherInputProps = CommonProps & {
@@ -220,6 +232,7 @@ class LabeledTextField extends React.Component<PropsWithForwardRef, State> {
             autoComplete,
             forwardedRef,
             ariaDescribedby,
+            name,
             // NOTE: We are not using this prop, but we need to remove it from
             // `otherProps` so it doesn't override the `handleValidate` function
             // call. We use `otherProps` due to a limitation in TypeScript where
@@ -267,6 +280,7 @@ class LabeledTextField extends React.Component<PropsWithForwardRef, State> {
                                 readOnly={readOnly}
                                 autoComplete={autoComplete}
                                 ref={forwardedRef}
+                                name={name}
                                 {...otherProps}
                             />
                         }